Move configs to a separate file
